Enable packaging for PyPi
