+++ MANIFEST.in
@@ -1,0 +1,3 @@
+include requirements.txt
+recursive-include static *
+recursive-include templates *
