+++ README
@@ -1,0 +1,24 @@
+Souma
+=====
+
+Souma is a peer to peer cognitive network, it allows communication and collaboration without the dangers and limitations of a centralized platform. By automatically coordinating group-forming processes it unleashes the Internet’s power of bringing together users with shared interests and intentions without the overhead of conventional institutions. Souma reveals itself as a magazine-like, paged website and can be accessed from every modern webbrowser once the client has been installed.
+
+Please see our Wiki for a more detailed description of what Souma does:
+
+https://github.com/ciex/souma/wiki/Souma-App-Overview
+
+Installation Guide
+==================
+
+1. `(sudo) python set_hosts.py` to add an entry for Souma to your /etc/hosts file
+2. `python run.py`
+
+Development Status
+==================
+
+Souma is currently in beta status. That means it should be expected to break and functionality is quite limited. Development is currently focussed on laying a groundwork in sensible protocols and APIs.
+
+License
+=======
+
+Souma is licensed under the Apache2 license. See LICENSE, NOTICE and the licenses/ folder for details.
